fixed cmakelist and others. not tested because of the caos of inverse_kinematics_node.cpp
