+++ CMakeLists.txt
@@ -6,30 +6,26 @@ if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
 endif()
 
 # find dependencies
-find_package(ament_cmake REQUIRED)
-# uncomment the following section in order to fill in
-# further dependencies manually.
-# find_package(<dependency> REQUIRED)
+find_package(ament_cmake_auto REQUIRED)
+ament_auto_find_build_dependencies()
 
-add_executable(inverse_kinematics_node src/inverse_kinematics_node.cpp)
-target_include_directories(inverse_kinematics_node PUBLIC
-  $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-  $<INSTALL_INTERFACE:include>)
-target_compile_features(inverse_kinematics_node PUBLIC c_std_99 cxx_std_17)  # Require C99 and C++17
+ament_auto_add_library(inverse_kinematics_node SHARED src/inverse_kinematics_node.cpp)
 
-install(TARGETS inverse_kinematics_node
-  DESTINATION lib/${PROJECT_NAME})
+install(TARGETS
+        inverse_kinematics_node
+        DESTINATION lib/${PROJECT_NAME}
+        LIBRARY DESTINATION lib)
 
 if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
   # the following line skips the linter which checks for copyrights
   # comment the line when a copyright and license is added to all source files
-  set(ament_cmake_copyright_FOUND TRUE)
+  #set(ament_cmake_copyright_FOUND TRUE)
   # the following line skips cpplint (only works in a git repo)
   # comment the line when this package is in a git repo and when
   # a copyright and license is added to all source files
-  set(ament_cmake_cpplint_FOUND TRUE)
+  #set(ament_cmake_cpplint_FOUND TRUE)
   ament_lint_auto_find_test_dependencies()
 endif()
 
-ament_package()
+ament_auto_package()
