--- cmake-build-debug/CMakeFiles/3.25.2/CMakeRCCompiler.cmake
@@ -1,6 +1,0 @@
-set(CMAKE_RC_COMPILER "C:/industry/CLion 2023.1.3/bin/mingw/bin/windres.exe")
-set(CMAKE_RC_COMPILER_ARG1 "")
-set(CMAKE_RC_COMPILER_LOADED 1)
-set(CMAKE_RC_SOURCE_FILE_EXTENSIONS rc;RC)
-set(CMAKE_RC_OUTPUT_EXTENSION .obj)
-set(CMAKE_RC_COMPILER_ENV_VAR "RC")
