--- cmake-build-debug/CMakeFiles/clion-Debug-log.txt
@@ -1,32 +1,0 @@
-"C:\industry\CLion 2023.1.3\bin\cmake\win\x64\bin\cmake.exe" -DCMAKE_BUILD_TYPE=Debug -G "CodeBlocks - MinGW Makefiles" -S \\wsl$\Ubuntu-22.04\home\yoshi\CatchRobo2023\ros2_ws\src\inverse_kinematics_package -B \\wsl$\Ubuntu-22.04\home\yoshi\CatchRobo2023\ros2_ws\src\inverse_kinematics_package\cmake-build-debug
--- The C compiler identification is GNU 11.2.0
--- The CXX compiler identification is GNU 11.2.0
--- Detecting C compiler ABI info
--- Detecting C compiler ABI info - done
--- Check for working C compiler: C:/industry/CLion 2023.1.3/bin/mingw/bin/gcc.exe - skipped
--- Detecting C compile features
--- Detecting C compile features - done
--- Detecting CXX compiler ABI info
--- Detecting CXX compiler ABI info - done
--- Check for working CXX compiler: C:/industry/CLion 2023.1.3/bin/mingw/bin/g++.exe - skipped
--- Detecting CXX compile features
--- Detecting CXX compile features - done
-CMake Error at CMakeLists.txt:9 (find_package):
-  By not providing "Findament_cmake.cmake" in CMAKE_MODULE_PATH this project
-  has asked CMake to find a package configuration file provided by
-  "ament_cmake", but CMake did not find one.
-
-  Could not find a package configuration file provided by "ament_cmake" with
-  any of the following names:
-
-    ament_cmakeConfig.cmake
-    ament_cmake-config.cmake
-
-  Add the installation prefix of "ament_cmake" to CMAKE_PREFIX_PATH or set
-  "ament_cmake_DIR" to a directory containing one of the above files.  If
-  "ament_cmake" provides a separate development package or SDK, be sure it
-  has been installed.
-
-
--- Configuring incomplete, errors occurred!
-See also "//wsl$/Ubuntu-22.04/home/yoshi/CatchRobo2023/ros2_ws/src/inverse_kinematics_package/cmake-build-debug/CMakeFiles/CMakeOutput.log".
