--- cmake-build-debug/CMakeFiles/clion-environment.txt
@@ -1,4 +1,0 @@
-ToolSet: w64 9.0 (local)@C:\industry\CLion 2023.1.3\bin\mingw
-Options: 
-
-Options:
