+++ src/inverse_kinematics_node.cpp
@@ -1,10 +1,10 @@
 #include <cstdio>
-#include "rclcpp/rclcpp.hpp"
+#include <rclcpp/rclcpp.hpp>
 #include <cmath>
 
 #include <memory>
 
-#include "robot_kinematics.h"
+#include "ros2_inverse_kinematics/robot_kinematics.h"
